RealEstate_app/requirements.txt
firecrawl-py
pandas
ollama
python-dotenv
openpyxl
langchain
langchain-community
llama-index-llms-groq 
streamlit